// src/types/cliente.ts
export interface Cliente {
    ID_CLIENTE?: number;
    NOMBRE: string;
    DIRECCION?: string;
    COD_POSTAL?: number;
    CIF?: string;
    TELEFONO?: string;
    FAX?: string;
    RESPONSABLE?: string;
    RESPONSABLE_OTROS?: string;
    RESPONSABLE_METROLOGIA?: string;
    CARGO?: string;
    TIPO?: string;
    EMAIL?: string;
    EMAIL2?: string;
    EMAIL_FACTURACION?: string;
    EMAIL_METROLOGIA?: string;
    EMAIL_ALODINE?: string;
    BANCO?: string;
    CUENTA?: string;
    OBSERVACIONES?: string;
    WEB?: string;
    PAIS_ID?: number;
    PROVINCIA_ID?: number;
    MUNICIPIO_ID?: number;
    ANULADO?: number;
    CENTRO?: string;
    FACTURA_DETERMINACIONES?: number;
    EADS?: number;
    AIRBUS?: number;
    IBERIA?: number;
    AGROALIMENTARIO?: number;
    EXTRANJERO?: number;
    INTRA?: number;
    IDIOMA_FACTURA?: number;
    FP_ID?: number;
    TARIFA_ID?: number;
    CLAVEWEB?: number;
    CC?: string;
    ENVIO_DIRECCION?: string;
    ENVIO_COD_POSTAL?: number;
    ENVIO_PAIS_ID?: number;
    ENVIO_PROVINCIA_ID?: number;
    ENVIO_MUNICIPIO_ID?: number;
    INFORMES_DIRECCION?: string;
    INFORMES_COD_POSTAL?: number;
    INFORMES_PAIS_ID?: number;
    INFORMES_PROVINCIA_ID?: number;
    INFORMES_MUNICIPIO_ID?: number;
    SECCION?: string;
    FACTURA_ELECTRONICA?: number;
    PLANT_ID?: number;
    PARENT_ID?: number;
    CALIBRY_ID?: number;
}